define/__init__.py
import os
from dotenv import load_dotenv

load_dotenv()

JWT_SECRET = os.getenv("JWT_SECRET", "kb-secret")
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
ELASTICSEARCH_URL = os.getenv("ELASTICSEARCH_URL", "http://127.0.0.1:9200")

