requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
elasticsearch==8.11.0
pyjwt==2.8.0
python-dotenv==1.0.0
openai==1.12.0
bcrypt==4.2.0

